refactored perfil
